+++ V2-Springboot/ALMXN/src/main/java/Project/ALMXN/models/Categoria.java
@@ -2,22 +2,22 @@ package Project.ALMXN.models;
 
 public class Categoria {
 
-    private Integer idCategoria;
+    private Long idCategoria;
     private String nombreCategoria;
     private String descripcionCategoria;
     private String estadoCategoria;
 
     public Categoria() {}
 
-    public Categoria(Integer idCategoria, String nombreCategoria, String descripcionCategoria, String estadoCategoria){
+    public Categoria(Long idCategoria, String nombreCategoria, String descripcionCategoria, String estadoCategoria){
         this.idCategoria = idCategoria;
         this.nombreCategoria = nombreCategoria;
         this.descripcionCategoria = descripcionCategoria;
         this.estadoCategoria = estadoCategoria;
     }
 
-    public Integer getIdCategoria() { return idCategoria;}
-    public void setIdCategoria(Integer idCategoria) { this.idCategoria = idCategoria;}
+    public Long getIdCategoria() { return idCategoria;}
+    public void setIdCategoria(Long idCategoria) { this.idCategoria = idCategoria;}
 
     public String getNombreCategoria() { return nombreCategoria; }
     public void setNombreCategoria(String nombreCategoria) { this.nombreCategoria = nombreCategoria; }
